ServiceSathi/backend/requirements.txt
cloudinary==1.44.2
Flask==3.1.3
protobuf==7.34.1
pyodbc==5.3.0
python-dotenv==1.2.2
Werkzeug==3.1.8
google-generativeai ==0.1.0
